+++ changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 2023-01-15
+
+### Added
+- add unique constraint to fields in Machine and Snack model
+- add signal to auto set status of Machine
+
+
 ## 2023-01-14
 
 ### Added